Tests/LibWeb/Text/expected/css/insert-rule-in-shadow-root-style-sheet.txt

before target: rgb(255, 0, 0)
before other: rgba(0, 0, 0, 0)
before host: rgb(0, 0, 0)
before pseudo: rgb(0, 0, 0)
before child selector: rgb(0, 0, 0)
after target: rgb(0, 128, 0)
after other: rgba(0, 0, 0, 0)
after host: rgb(0, 0, 255)
after pseudo: rgb(255, 0, 0)
after child selector: rgb(128, 0, 128)
